requirements.worker.txt
requests>=2.31,<3.0
paho-mqtt>=1.6,<2.0
python-dotenv>=1.0,<2.0
langchain>=0.2,<0.4
langchain-core>=0.2,<0.4
langchain-openai>=0.1,<0.3
